Convert in-place pubsub scripts to async execution
This commit allows lua scripts subscribed to a pubsub
channel to PUBLISH to other channels that also have
listening lua scripts.

Yes, if you do The Bad Thing and subscribe a channel to
itself you'll infinite loop Redis.  Do not infinite loop Redis.
(Note: this also allows an indirect way for lua scripts to
call other lua scripts)

How it works: Each PUBLISH to a channel with a listening
lua script gets appended to a list of events to process
during the next server cron (default: 10 Hz).

Before this commit, lua scripts were executed before the PUBLISH
command returned to the caller.  Now lua scripts get executed on
average 50ms after your PUBLISH command.

+++ src/scripting.c
@@ -1022,7 +1022,7 @@ void evalName(redisClient *c, char *name) {
    because of how we initialize fields below.  If you find
    a need for more arguments against a script, feel free to
    increase max_args below. */
-void evalNameWithArgs(redisClient *c, char *name, int argc, ...) {
+void evalNameWithArgs(redisClient *c, char *name, long argc, ...) {
     va_list ap;
     int total_args = argc + 3;
     int j;
